# user_manage/api-docs.yml
openapi: 3.0.0
info:
  title: User Service API
  version: "1.0"
paths:
  /node/users:
    get:
      tags:
        - Users
      summary: Returns a list of users
      description: Retrieve a list of users from the system
      responses:
        '200':
          description: A JSON array of user objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
  /node/send:
    post:
      summary: Send a message
      description: Sends a message to RabbitMQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: "Hello, RabbitMQ!"
      responses:
        "200":
          description: Message successfully sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message sent to RabbitMQ"
        "400":
          description: Bad request - Message is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Message is required"
        "500":
          description: Internal Server Error - Error sending message to RabbitMQ
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error sending message to RabbitMQ"
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user by ID
      description: Retrieve a single user by their unique identifier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique identifier of the user to retrieve
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Internal server error

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string